+++ .gitignore
@@ -1,3 +1,3 @@
 __pycache__/
 test_inputs/
-
+data/
